+++ docs/conf.py
@@ -208,8 +208,8 @@ rst_prolog = """
 .. |DetectorReader| replace:: :class:`~serpentTools.DetectorReader`
 .. |DepletionReader| replace:: :class:`~serpentTools.DepletionReader`
 .. |DepmtxReader| replace:: :class:`~serpentTools.DepmtxReader`
-.. |BranchCollector| replace:: :class:`~serpentTools.BranchCollector`
-.. |BranchedUniv| replace:: :class:`~serpentTools.BranchedUniv`
+.. |BranchCollector| replace:: :class:`~serpentTools.xs.BranchCollector`
+.. |BranchedUniv| replace:: :class:`~serpentTools.xs.BranchedUniv`
 .. |BranchingReader| replace:: :class:`~serpentTools.BranchingReader`
 .. |XSPlotReader| replace:: :class:`~serpentTools.XSPlotReader`
 .. |HistoryReader| replace:: :class:`~serpentTools.HistoryReader`